pyproject.toml
[project]
name = "cafebert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.5.0",
    "fire>=0.7.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "torch>=2.6.0",
    "transformers>=4.51.2",
]
